python/requirements.txt
altgraph==0.17.4
bitarray==2.9.2
bitstring==4.2.3
cffi==1.16.0
cryptography==42.0.7
ecdsa==0.19.0
esptool==4.7.0
gitdb==4.0.11
GitPython==3.1.43
importlib_metadata==7.1.0
intelhex==2.3.0
packaging==24.0
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyserial==3.5
pywin32-ctypes==0.2.2
PyYAML==6.0.1
reedsolo==1.7.0
six==1.16.0
smmap==5.0.1
zipp==3.19.1
